Fix for log rotation and some cleanups.
If the log file is smaller then expected set the LogSucker position to
0.
If the client reports that it isn't playing something any more cancel
the watching (only works on some clients.
Change warning and fix to debug logging.
Only set last used action to something if we get a good response from
Trakt, this will hopefully fix network errors.